refactor: createViewController 메서드 내 ViewController Type별 분기 #3
- TypeOfViewController의 expositionItem 연관값 추가하여 데이터를 전달
- 관련 Namespace 삭제
- ViewControllerFactory 파일 위치 변경
