+++ Expo1900/Expo1900/Controller/ExpositionItemTableViewController.swift
@@ -69,11 +69,7 @@ extension ExpositionItemTableViewController {
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let selectedItem = self.expositionItems[indexPath.row]
         
-        let expositionItemStoryboard = UIStoryboard(name: StoryboardFileName.expositionItem, bundle: nil)
-
-        let expositionItemViewController = expositionItemStoryboard.instantiateViewController(identifier: StoryboardIdentifierName.expositionItem) { coder in
-            return ExpositionItemViewController(coder: coder, expositionItem: selectedItem)
-        }
+        let expositionItemViewController = ViewControllerFactory.createViewController(of: .expositionItem(data: selectedItem))
 
         self.navigationController?.pushViewController(expositionItemViewController, animated: true)
     }
